refactor(sandbox): move lifecycle reconciliation to orchestration

+++ internal/workeragent/command.go
@@ -147,5 +147,5 @@ func availableStorageBytes(path string) int64 {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0
 	}
-	return int64(stat.Bavail) * int64(stat.Bsize)
+	return int64(stat.Bavail) * stat.Bsize
 }
